Drop support for BME280 in favor of stability
During extensive testing we noticed that with the current firmware we
cannot read the I2C sensors after some minutes. Potentially the bus locks
up and you can only resolve this by powering off the board for some minutes.
A simple reset does NOT do the job.

I am afraid that there is a clash between the two BME drivers.

To overcome this, this PR removes support for BME280. With my test board I
could get the sensor data for days without a lock up.